requirements.txt
requests
beautifulsoup4
lxml
schedule
flask
werkzeug
jinja2
markupsafe
click
itsdangerous
blinker
supabase
python-dotenv
flask-login
google-auth
google-auth-oauthlib
google-auth-httplib2